📦 chore(layout.tsx): add new layout component for board/[boardId] route 📦 chore(page.tsx): add new page component for board/[boardId] route
The layout.tsx file adds a new layout component for the board/[boardId] route. This component is responsible for rendering the layout of the board page, including the background image and the main content area.

The page.tsx file adds a new page component for the board/[boardId] route. This component represents the actual content of the board page and currently displays a placeholder text "Board id page".